Clear refactor of backend code
