+++ backend/Models/Cell.cs
